pre_process

This function preprocess the data-inputs and automatically sets uniform format for future use.

Usage

pre_process(
  dataList,
  typenameList = NULL,
  replaceNA = TRUE,
  scale = TRUE,
  autoColName = "Day_"
)

Arguments

dataList

A list of datasets. Note that for each dataset, the first columns should be ID and the rest columns are Time1, Time2, ...

typenameList

A vector of dataset names.

scale

Logical; if TRUE (default), scale the datasets by row. See also scale_by_row__

autoColName

A string; if it's not NULL (default), automatically set uniform column names for all the datasets.

replaceNa

Logical; if TRUE (default), replace NA with 0.

Examples

data(samples)
a <- pre_process(samples)
head(a,10)
#>                 ID       Day_1        Day_2      Day_3       Day_4      Day_5
#> 105494 grp1_123656  1.16324449  1.363903662  1.3037049  0.70082791  0.2386620
#> 101572 grp1_118996  2.22195307  0.921244789 -0.6520548 -0.47274459  0.7920261
#> 122928 grp1_144232  2.21925785 -1.355860458 -0.5660429 -0.09238685  0.5524392
#> 65357   grp1_76657  1.66012361 -0.199676781 -0.9443640  0.39394648  1.6731019
#> 112264 grp1_131716 -1.93475260 -1.747281361 -0.5416194  0.06701984  0.5028177
#> 49395   grp1_58305  0.07723733 -0.003775164  1.6222273 -1.35291009 -1.2707773
#> 48371   grp1_57166 -1.10393923  2.019796561  1.7547445 -0.58148212 -0.4948189
#> 137473 grp1_160925 -1.42596862  0.921180593  1.1847738  0.20269795 -0.6186228
#> 20907   grp1_24862 -1.38706943  0.139524447  1.2329288 -0.97446897 -1.1633956
#> 37886   grp1_44863  0.10658683 -0.020718460 -0.7292620 -0.16404407  2.0150476
#>              Day_6      Day_7       Day_8      Day_9     Day_10      Day_11
#> 105494 -0.51320765 -0.4941146 -1.16197630 -1.0444529 -1.2134812 -0.34311029
#> 101572 -1.26690762 -0.4369331  0.29679523 -0.3799502 -0.9511152 -0.07231373
#> 122928  0.16510415  0.8255533 -0.81083343  0.0613068  0.1064795 -1.10501724
#> 65357  -0.09385111  0.1260783 -0.25759942 -0.1881348 -0.4223389 -1.74728521
#> 112264  0.63638291  0.8979775  0.70722909  0.8780265  0.1489132  0.38528667
#> 49395   0.14045162  0.3075112  0.28800328  0.7235652 -1.5017100  0.97017677
#> 48371  -0.84588429 -0.2203122 -0.09852822 -0.4954710 -0.0624898  0.12838479
#> 137473 -1.09819122 -0.4719112 -0.49017273  0.1452272 -0.2322416  1.88322854
#> 20907  -0.30876062  0.3860328  0.46505907 -0.5159709  1.8524665  0.27365381
#> 37886   1.14703159  0.7834282 -0.03173524 -1.2593593 -1.1945658 -0.65240944
#>             type
#> 105494 Dataset_1
#> 101572 Dataset_1
#> 122928 Dataset_1
#> 65357  Dataset_1
#> 112264 Dataset_1
#> 49395  Dataset_1
#> 48371  Dataset_1
#> 137473 Dataset_1
#> 20907  Dataset_1
#> 37886  Dataset_1