transpose dataframe

This function transpose a given data frame, converting the first column to new column names

Usage

transpose_dataframe__(data)

Format

A data frame in which the first column is ID and the rest are observations for different time points.

Examples

data(fungal)
transpose_dataframe__(fungal[[3]])
#>        grp3_1     grp3_2     grp3_3       grp3_4    grp3_5     grp3_6
#> 1  -0.4702515 -1.6644934 -0.9248744  0.047325767 -1.301994 -1.2295733
#> 2  -0.3390523 -2.6212720 -0.6240496  0.265735306 -1.229585 -0.8622896
#> 3  -0.5013460 -2.2834156 -0.5520060  0.201532312 -1.302000 -0.9248781
#> 4  -0.4833872 -1.8775744 -0.2821400  0.220536581 -1.440303 -1.1003985
#> 5  -0.4902185 -0.6482506 -0.3161406  0.090037856 -1.552596 -1.0853967
#> 6  -0.6680519 -1.5863859 -0.3980750 -0.062929298 -1.747046 -1.3012848
#> 7  -0.6609125 -1.7470607 -0.3106139 -0.127386629 -1.695266 -1.1815061
#> 8  -0.7754728 -0.3588664 -0.2878704  0.009608468 -1.747025 -1.2580652
#> 9  -0.7022775 -1.0980109 -0.2226940 -0.042702042 -1.775274 -1.2724684
#> 10 -0.5814545 -1.7763473 -0.1963857  0.085840662 -1.588424 -1.1067230
#> 11 -0.1418371 -0.4198883  0.4274527  0.312484614 -1.296980 -0.9024968
#>        grp3_7     grp3_8     grp3_9   grp3_10     grp3_11     grp3_12
#> 1  -2.1189770 -0.9483767 -0.1387409 0.8102653  0.69515714  0.49145477
#> 2  -1.0440039 -0.1836356  0.3972186 0.8486148  0.61005827  0.15349866
#> 3  -1.1183906 -0.1616758  0.4116574 0.8923234  0.42987339 -0.03997297
#> 4  -0.8186625  0.1492755  0.5675773 0.7734366  0.37725017 -0.18687746
#> 5  -0.7422817  0.1706407  0.6422367 0.7640935  0.42485893 -0.02766762
#> 6  -0.8391938  0.2129877  0.7022987 0.7800061  0.27844081 -0.24693878
#> 7  -0.7616431  0.2220469  0.6546750 0.7222672  0.09072649 -0.21396653
#> 8  -0.7294594  0.1713840  0.7120832 0.6546163 -0.10946565 -0.34255210
#> 9  -0.7294585  0.2685916  0.7607575 0.6927528  0.05771203 -0.26807681
#> 10 -0.5661721  0.3819543  0.7544223 0.6100550  0.04115516 -0.20312891
#> 11  0.2163800  0.7608816  1.0977554 0.7394590  0.34197613  0.10704836
#>       grp3_13   grp3_14   grp3_15     grp3_16     grp3_17    grp3_18    grp3_19
#> 1  -0.3701520 -3.077605 -1.371815  0.48569778  0.27503630 -1.0596873 -0.8943071
#> 2  -0.6914459 -3.077622 -1.588434  0.50517436  0.26999617 -0.8752052 -1.0596985
#> 3  -0.8858472 -2.930967 -1.876307  0.25125700  0.04116377 -0.8243861 -0.9409440
#> 4  -0.9836278 -2.931589 -2.059688  0.17064051 -0.20634798 -0.7242997 -1.1841799
#> 5  -0.8548463 -2.931593 -2.059691 -0.01435771 -0.23100914 -0.5422444 -1.2185383
#> 6  -0.9907782 -3.077485 -2.055217 -0.21397482 -0.50605888 -0.2878558 -1.1922453
#> 7  -1.0215814 -3.077519 -2.117582 -0.11917240 -0.51240222 -0.4363825 -1.2580816
#> 8  -0.9907757 -3.077431 -2.055189 -0.03740912 -0.29297686 -0.4103282 -1.2024605
#> 9  -0.9320405 -2.930322 -2.055205 -0.09126199 -0.49896360 -0.3864553 -1.3534197
#> 10 -0.7887667 -2.930953 -1.929570  0.07663091 -0.28982775 -0.4532874 -1.1710357
#> 11 -0.4382779 -3.024819 -1.942858 -0.07994823 -0.32392776 -0.1018753 -1.0578820
#>        grp3_20   grp3_21     grp3_22     grp3_23    grp3_24     grp3_25
#> 1  -0.29509360 0.7145905  0.17954670 -1.07229062 -1.8390608 -0.61875477
#> 2  -0.12300779 0.7544264 -0.15420099 -0.70596119 -1.7763583 -0.45329050
#> 3  -0.26514011 0.6486077 -0.39465324 -0.61260184 -1.5884280 -0.44699942
#> 4   0.06372819 0.6879448 -0.37815200 -0.46383129 -1.3559718 -0.40255742
#> 5   0.18147138 0.8644581 -0.17327861 -0.21271335 -1.0451223 -0.27707149
#> 6   0.26353692 0.8587878 -0.19390302  0.18213036 -0.7294556 -0.19988255
#> 7   0.16416015 0.7519810 -0.25703397  0.07222459 -0.8604706 -0.45818059
#> 8   0.26350894 0.7691569 -0.24002166  0.22200516 -0.7022787 -0.50606814
#> 9   0.45057790 0.8235551 -0.28310640  0.16413319 -0.7754725 -0.35886070
#> 10  0.36747832 0.8616172 -0.19042570  0.09510733 -0.8321374 -0.38201796
#> 11  1.01234929 1.3255373  0.06331094  0.32725659 -0.5256676  0.02886566
#>       grp3_26    grp3_27    grp3_28    grp3_29     grp3_30    grp3_31   grp3_32
#> 1  -1.6966886 -1.6151556 -0.4801848 -0.9409402 -0.57364257 -0.7958473 0.4657598
#> 2  -1.2036074 -1.1560362 -0.6306802 -1.1183974 -0.32569792 -0.8858544 0.5939579
#> 3  -0.9483805 -0.9924673 -0.6378357 -1.1934464 -0.42481558 -0.7887645 0.4769347
#> 4  -0.8428663 -0.9496579 -0.5939028 -0.9746419 -0.15807829 -0.6723695 0.6508907
#> 5  -0.5836166 -0.6325320 -0.3287612 -0.8958372  0.04431516 -0.7243007 0.8118830
#> 6  -0.2984656 -0.3798758 -0.2929624 -0.6609141  0.29114544 -0.9470994 0.7992809
#> 7  -0.4292262 -0.4672282 -0.4103112 -0.6359809  0.21300236 -0.9567560 0.7894116
#> 8  -0.2570569 -0.3172101 -0.3106348 -0.4769818  0.40910613 -1.0122861 0.7992399
#> 9  -0.2984734 -0.4225975 -0.2570499 -0.3980817  0.42105218 -0.8145494 0.7691774
#> 10 -0.3892175 -0.4390607 -0.2801241 -0.4610189  0.37488162 -0.9409451 0.8018439
#> 11  0.1544571 -0.1880659  0.2371302  0.6520338  1.21130930 -0.2799783 1.2563233
#>        grp3_33    grp3_34    grp3_35   grp3_36   grp3_37     grp3_38
#> 1   0.30064508 -0.4187891 -0.1904186 0.7826672 -1.749313 -0.70595102
#> 2   0.11501536 -0.8166162  0.2167651 0.8871536 -2.283420 -0.20313050
#> 3  -0.01043304 -1.0068630  0.2939139 0.7544363 -2.118986 -0.04627619
#> 4   0.13875452 -0.8548450  0.3618760 0.7853186 -1.751585  0.28226834
#> 5   0.39289956 -0.4638319  0.3995325 0.6718302 -1.777422  0.16330664
#> 6   0.43527222 -0.4859102  0.6073689 0.6619914 -1.927873  0.16414621
#> 7   0.24798867 -0.6172441  0.4956702 0.5859071 -1.875052  0.06417958
#> 8   0.18732763 -0.6532130  0.4430148 0.5178590 -1.836685  0.10016919
#> 9   0.23876184 -0.7211749  0.5375136 0.4956346 -1.970067  0.32636233
#> 10  0.20836379 -0.6557337  0.4856887 0.5126184 -1.839063  0.11501504
#> 11  0.53092965 -0.2926948  1.0646341 0.8887575 -1.397793  0.35862115
#>         grp3_39     grp3_40     grp3_41    grp3_42   grp3_43     grp3_44
#> 1   0.195813562 -0.97229254 -1.09065301 -0.9132183 0.8486202 -0.59109352
#> 2   0.067907271 -0.54016530 -1.49833239 -1.0992138 0.9249664 -0.67022244
#> 3  -0.226856477 -0.33904420 -0.81660888 -1.2170490 0.9491440 -0.60352250
#> 4  -0.340902953 -0.22115418 -0.44957482 -1.5904619 0.9395119 -0.62540708
#> 5   0.156924280 -0.06832951  0.59588132 -1.4403046 1.0363814 -0.39611338
#> 6  -0.027758837  0.05041795  0.17140977 -1.6138056 0.9950895 -0.33055715
#> 7  -0.165079606  0.12133987  0.04455634 -1.6620926 1.0097971 -0.36729635
#> 8   0.314467056  0.16412064  0.96970062 -1.3896174 1.0437496  0.29807922
#> 9  -0.006496803  0.17139667  0.70227964 -1.5252498 1.0308830 -0.02144081
#> 10 -0.253827026  0.18432484 -0.17673867 -1.5269407 1.0496552 -0.49421256
#> 11  0.671477435  0.63262152  1.74740483 -0.9989334 2.0071460  0.47422572
#>     grp3_45     grp3_46    grp3_47    grp3_48    grp3_49   grp3_50     grp3_51
#> 1  1.615784  0.93096745 -0.3390432 -0.5520040 -2.1733895 0.9913193 -0.16716683
#> 2  1.671094  0.68643848 -0.3701612  0.0766310 -1.3718275 0.9343724 -0.37630895
#> 3  1.518976  0.58134320 -0.3763009 -0.3332386 -1.1710364 0.8743370 -0.52060342
#> 4  1.406476  0.32505606 -0.7338682  0.2679363 -0.9603508 0.5958805 -0.68288844
#> 5  1.341878  0.26250943 -0.8186636  0.8767849 -1.1841816 0.6127494 -0.50979602
#> 6  1.317413  0.06416713 -0.5507966  0.6167565 -1.0822318 0.7415360 -0.51881225
#> 7  1.231841 -0.04268050 -0.7487513  0.8588121 -0.9699445 0.6768601 -0.55664894
#> 8  1.270688 -0.15030748 -0.7211759  1.0807818 -0.8604699 0.6849014 -0.62267914
#> 9  1.251895 -0.11919193 -0.6887263  0.8845991 -0.8501162 0.6546352 -0.71384638
#> 10 1.254201 -0.07831837 -0.8166109  0.9309570 -1.0439973 0.6864347 -0.50793484
#> 11 2.456646  0.19785041 -0.1230100  2.1764858 -0.6556166 1.1706002 -0.06320768
#>       grp3_52   grp3_53   grp3_54     grp3_55    grp3_56     grp3_57
#> 1  -0.2801173 -2.283406 1.0533973  0.08584858 -0.2099907  0.27000338
#> 2  -0.5013539 -1.664506 1.2053331  0.03332014  0.1742352  0.17953921
#> 3  -0.5736446 -1.283230 1.1097133 -0.24309134  0.2592622  0.15350757
#> 4  -0.9070225 -1.062103 0.9319973 -0.56121451  0.2115731 -0.42703161
#> 5  -0.9603522 -1.194646 0.8503965 -0.76769641  0.3134276  0.02860182
#> 6  -0.7211719 -1.181500 0.7121221 -1.01228913  0.4273436 -0.27808853
#> 7  -0.6105601 -1.292972 0.6463295 -1.08980808  0.3576907 -0.30420462
#> 8  -0.6359890 -1.243604 0.5710846 -1.08222700  0.3434789  0.77996551
#> 9  -0.7616458 -1.105752 0.5711025 -1.15348078  0.4017708  0.39489203
#> 10 -0.6306761 -1.273869 0.5398420 -1.20359980  0.4229587 -0.43145747
#> 11 -0.3749138 -1.078708 1.0325898 -0.80673068  0.9286939  1.43695128
#>        grp3_58   grp3_59     grp3_60   grp3_61    grp3_62     grp3_63
#> 1  -0.67021236 -1.588421 -0.01788518 1.1351602 -0.1188759 -0.41295730
#> 2  -0.31890464 -1.929582 -0.03255790 1.0533925 -0.4880759 -0.52758536
#> 3  -0.28011794 -1.615163 -0.05969970 0.8396808 -0.6306726 -0.36198510
#> 4  -0.55321222 -1.316407 -0.05646554 0.7170496 -0.8877263 -0.06273235
#> 5  -0.40787623 -1.420066 -0.11857489 0.6314267 -0.9347703 -0.42703222
#> 6  -0.36730235 -1.370489 -0.01434460 0.5920404 -0.6226727 -0.53194547
#> 7  -0.39806948 -1.416179  0.15763415 0.5272699 -0.9396312 -0.53194193
#> 8  -0.26808372 -1.282365  0.13139180 0.4505616 -0.7487550 -0.55080479
#> 9  -0.31062840 -1.314037  0.15005607 0.5178762 -0.7867502 -0.36730938
#> 10 -0.25952990 -1.391069  0.18999145 0.5466685 -0.8751996 -0.62404545
#> 11 -0.01312118 -1.205507  0.58719554 0.8319846 -0.3386473  0.39514285
#>       grp3_64    grp3_65   grp3_66    grp3_67     grp3_68    grp3_69
#> 1  -1.0992025 0.01447009 0.8298921 -0.1188759 -0.68032775 1.09398923
#> 2  -0.7887718 0.18999221 0.8396707 -0.1140297 -0.91322910 0.98538320
#> 3  -0.5910957 0.22587828 0.7826715 -0.1140212 -0.85248462 1.01824147
#> 4  -0.2127130 0.34160747 0.8983387 -0.2461638 -1.02499221 0.87678368
#> 5  -0.3781525 0.34877182 0.9603893  0.1770598 -0.39196512 0.66544677
#> 6  -0.2072579 0.51789321 0.9831953  0.1934172 -0.64387667 0.28382280
#> 7  -0.1696664 0.62612644 0.9951158  0.2635524 -0.27250156 0.20516774
#> 8  -0.1803803 0.63642760 1.0342155  0.2685777 -0.09127121 0.07218869
#> 9  -0.1094567 0.71210287 1.0308830  0.2725933 -0.33056459 0.23093905
#> 10 -0.3530979 0.67017830 1.0124127  0.2803512 -0.17148690 0.12600048
#> 11  0.4115484 1.80939791 1.9033850  0.4900500  0.50804773 0.86167104
#>         grp3_70  grp3_71   grp3_72   grp3_73    grp3_74     grp3_75    grp3_76
#> 1  -0.161675715 1.404977 0.9721990 -2.930954  0.3461468 -1.34675473 -0.7227356
#> 2  -0.138749359 1.367733 0.6951512 -2.930971  0.3534122 -0.80058875 -0.4314604
#> 3   0.115024213 1.318745 0.6010053 -3.077619  0.2340562 -1.10306728 -0.4610141
#> 4   0.130664091 1.095482 0.3928990 -3.077754 -0.3359091 -0.63610650 -0.6070717
#> 5   0.009458161 1.070477 0.4698211 -3.077758 -0.5612153  0.07634078 -0.4960247
#> 6   0.009630276 1.121635 0.4430472 -2.481041 -0.9699415 -0.11437480 -0.4672327
#> 7  -0.135025664 1.080857 0.2686213 -2.500260 -1.0122928  0.23470739 -0.4769675
#> 8  -0.101960566 1.009724 0.1213020 -2.453128 -0.9470979  0.47879129 -0.5566601
#> 9   0.064155239 1.092472 0.1314039 -2.438192 -1.0706203  0.35440595 -0.5380686
#> 10 -0.138748159 1.035310 0.2426621 -3.077603 -1.0722951  0.33714968 -0.4060449
#> 11  0.300404582 1.694798 0.4469993 -2.642191 -0.5579436  1.56737091 -0.2040701
#>       grp3_77    grp3_78    grp3_79    grp3_80    grp3_81   grp3_82     grp3_83
#> 1   0.2426705 -0.7887614  0.1044158 -0.8321322 -2.6212560 -1.618898 -1.34675473
#> 2  -0.2849627 -1.0906644 -0.5814584 -1.1828488 -2.1734042 -2.041645 -0.80058875
#> 3  -0.2849545 -0.9132219 -0.7316662 -1.3910713 -2.4310584 -2.063123 -1.10306728
#> 4  -0.4336651 -0.7972499 -0.8778243 -1.9725078 -1.8414197 -2.622946 -0.63610650
#> 5  -0.8337415 -0.7815120 -0.6723705 -2.1204082 -1.1373955 -2.285332  0.07634078
#> 6  -1.2155562 -0.7867479 -0.8145474 -2.4302946 -1.4625157 -2.117562 -0.11437480
#> 7  -1.3362602 -0.9029027 -0.8839629 -2.9303763 -1.2155629 -1.970096  0.23470739
#> 8  -1.6620603 -0.9699395 -0.9806837 -2.9302943 -0.9320388 -1.875011  0.47879129
#> 9  -1.4161642 -0.9699417 -0.9567535 -3.0774611 -1.0428694 -2.117549  0.35440595
#> 10 -1.6151585 -1.0130205 -0.9049669 -2.7498847 -1.0906575 -2.118978  0.33714968
#> 11 -0.8536217 -0.7283125 -0.5782391 -2.1123892 -0.1601422 -1.175633  1.56737091
#>       grp3_84    grp3_85    grp3_86    grp3_87    grp3_88   grp3_89   grp3_90
#> 1  -0.3820114 -1.4389056 -0.4610126 -0.9924633 -0.5144436 0.7438966 0.8205633
#> 2  -0.6460750 -1.4641983 -1.0232949 -1.3910776 -0.8524919 0.6485979 0.6009957
#> 3  -0.7784970 -1.3152252 -1.0722950 -1.4181202 -0.9821593 0.5308581 0.4043220
#> 4   0.5834894 -1.4658621 -0.7676953 -1.2468742 -1.1373938 0.5485864 0.4929521
#> 5  -0.8088180 -1.5004727 -1.0137513 -1.4658642 -1.0915117 0.5675781 0.5570723
#> 6  -0.8927803 -1.2724716 -1.0428700 -1.4375163 -1.0215776 0.4956513 0.3870275
#> 7  -0.7294550 -1.4961840 -0.9806892 -1.5498449 -1.0428741 0.5103276 0.4210859
#> 8  -0.7944488 -1.3704768 -1.1814927 -1.4961583 -0.9118273 0.5614217 0.3319007
#> 9  -0.8839623 -1.3896251 -1.2436095 -1.2155537 -1.2580712 0.3870119 0.3793642
#> 10 -0.7227411 -1.4641891 -1.0838162 -1.3152234 -0.8410353 0.5874660 0.4914457
#> 11 -0.6174550 -0.8266778 -0.7822849 -1.0248793 -0.6811350 0.9881659 0.7852677
#>       grp3_91     grp3_92     grp3_93   grp3_94    grp3_95   grp3_96    grp3_97
#> 1  -0.6035203 -2.37055724  0.14454582 0.9913193 -0.5401556 0.5939642 -0.2031218
#> 2  -1.1067300 -2.11899144  0.01446210 0.7826616 -0.7403701 0.3674802 -0.6378433
#> 3  -1.3546998 -1.92957669 -0.17147992 0.7046084 -0.7157465 0.1795482 -0.8410334
#> 4  -1.2308816 -1.17271139 -0.19884516 0.7926453 -0.7422806 0.2465442 -0.7815109
#> 5  -1.2047317 -0.57578736 -0.09061836 0.8983400 -0.6254080 0.3303734 -0.6396828
#> 6  -1.2580742 -0.58827670 -0.16508863 0.8280235 -0.6359829 0.3144963 -0.7944466
#> 7  -1.2024747 -0.57927441 -0.28309116 0.7023207 -0.7211712 0.3378388 -0.8305260
#> 8  -1.1693484  0.37248816 -0.22270151 0.7222058 -0.6438826 0.2309256 -0.8305265
#> 9  -1.2282685  0.05040622 -0.28786377 0.6167384 -0.8305269 0.3043135 -0.6777652
#> 10 -1.0992069 -0.75139580 -0.11402860 0.8102552 -0.8943121 0.4539140 -0.5910994
#> 11 -0.6363155  0.91076413  0.17114281 1.3784615  0.0844919 0.7155045 -0.2424877
#>        grp3_98    grp3_99   grp3_100  grp3_101   grp3_102  grp3_103  grp3_104
#> 1   0.41165486 -1.8762987 -0.8243829 0.8396762 0.58133985 1.0353206 -1.203596
#> 2   0.14453822 -0.3892202 -1.4389180 0.5645177 0.28035250 1.1234492 -1.749326
#> 3  -0.20312217 -0.8063804 -1.4641920 0.5939676 0.05392313 0.8616320 -1.664501
#> 4  -0.13435199 -0.2559521 -1.4200636 0.6215665 0.20399964 1.0150685 -1.552594
#> 5  -0.02140940  0.3082798 -1.1004001 0.7926465 0.29666404 0.9875771 -1.751588
#> 6  -0.09125186  0.2387757 -1.3140402 0.5990212 0.15762031 0.8718729 -1.875035
#> 7  -0.07374226  0.5179124 -1.3140484 0.5520958 0.18736805 0.8718974 -1.927890
#> 8  -0.15874884  0.6685034 -1.5252403 0.3643107 0.01947366 0.8424528 -1.927848
#> 9  -0.07376298  0.5102915 -1.5863792 0.2559801 0.07220004 0.9571406 -1.747038
#> 10  0.04731796  0.6637119 -1.4983231 0.4657508 0.16046335 1.0681207 -1.749315
#> 11  0.37690967  1.6367123 -1.1042250 0.8091350 0.54950196 1.4977875 -1.268665
#>      grp3_105   grp3_106     grp3_107   grp3_108    grp3_109   grp3_110
#> 1  -1.1560247 -0.8524813 -2.749885110 -1.1934414  0.18999967 -0.9049619
#> 2  -1.3382039 -1.2452528 -1.293985702 -1.2738770  0.31144912 -1.1934531
#> 3  -1.2939792 -1.4983262 -1.839068822 -1.0906575  0.30631096 -1.1828420
#> 4  -1.3925080 -1.3731493 -1.119947453 -1.0375480  0.18147112 -1.2949915
#> 5  -1.3164085 -1.3027153 -0.553213009 -1.0375495  0.09949192 -1.2468759
#> 6  -1.1168283 -1.3362517 -0.538063538 -0.7487516  0.17761108 -1.2282711
#> 7  -1.2436194 -1.3534326 -0.278081199 -0.8725622  0.05043020 -1.5252671
#> 8  -1.1330866 -1.4375029 -0.006507126 -0.8228848 -0.01436583 -1.5498171
#> 9  -1.1330909 -1.3704842 -0.169684631 -0.8391955  0.00961900 -1.4961674
#> 10 -1.1934456 -1.2295773 -0.046284137 -0.8243881  0.02405433 -1.4389090
#> 11 -0.4784097 -0.8786460  1.130888465 -0.4948746  0.25758704 -1.1481677
#>      grp3_111   grp3_112   grp3_113    grp3_114  grp3_115   grp3_116  grp3_117
#> 1  -1.0130158 -0.6460650 -1.9295683 -1.13524280 0.2015307 -1.0232838 1.0829415
#> 2  -1.0723019 -0.9248852 -1.3020061 -0.65573798 0.3283731 -1.0068700 1.2053331
#> 3  -0.8943106 -0.9722965 -1.0232880 -0.47025311 0.3223723 -0.9585553 1.0400805
#> 4  -0.7907735 -1.2047300 -0.9271023 -0.31614019 0.4470618 -1.1076907 1.0297030
#> 5  -0.7540250 -1.1585673 -0.9746433 -0.28214038 0.4409785 -1.0621046 1.0150699
#> 6  -0.5253633 -1.5252558 -0.5999575 -0.03738846 0.5791814 -0.8725606 0.9299378
#> 7  -0.6777603 -1.4375263 -0.6288053 -0.19389443 0.6073894 -0.7754702 0.9059783
#> 8  -0.6777667 -1.6137879 -0.5253721 -0.08327323 0.7022601 -0.6680572 0.8997138
#> 9  -0.6532111 -1.8750251 -0.5319506 -0.08326391 0.6849207 -0.6438806 0.9154773
#> 10 -0.8858487 -1.5512124 -0.6622778 -0.08695166 0.7145807 -0.7403653 0.9381345
#> 11 -0.5102223 -0.9744216 -0.1716585  0.28416213 1.2950901 -0.2674806 1.4705949
#>     grp3_118   grp3_119  grp3_120   grp3_121   grp3_122    grp3_123   grp3_124
#> 1  0.6637215 -0.7403598 -1.526938 -1.4641858 -0.5814486  0.07663880 -1.1828371
#> 2  0.4769253 -1.1352543 -1.839074 -1.2170556 -0.7227458  0.23404714 -0.9409511
#> 3  0.3114583 -1.2594732 -1.551216 -1.0838162 -0.6914384  0.24267227 -0.7513932
#> 4  0.3134272 -1.2608675 -1.698135 -0.8337403 -1.0853951 -0.00654606 -0.8258841
#> 5  0.3772507 -1.1727131 -1.616517 -0.8640902 -1.2308833 -0.13435218 -0.8877275
#> 6  0.3263772 -1.1693553 -1.389630 -0.4923857 -1.1534826 -0.05665439 -0.6288074
#> 7  0.1990624 -1.1693613 -1.301293 -0.6532056 -1.2724792 -0.29845855 -0.8039346
#> 8  0.2051266 -1.1057481 -1.272462 -0.3744558 -1.1168227 -0.19989926 -0.7384412
#> 9  0.2838084 -1.2929608 -1.336248 -0.4514862 -1.0215772 -0.23226143 -0.7384404
#> 10 0.3283715 -1.0356846 -1.293978 -0.3763063 -1.0596918 -0.18363417 -0.7646783
#> 11 0.6080880 -0.6034163 -1.241273  0.2695784 -0.3596245  0.13267775 -0.2554560
#>      grp3_125    grp3_126   grp3_127   grp3_128   grp3_129  grp3_130  grp3_131
#> 1  -1.0356800  0.03332807 -0.9821553 -0.6557280 -0.5206016 -1.971286 0.3534192
#> 2  -0.5079382 -0.01789327 -1.5512221 -0.5520137 -0.6126095 -2.057463 0.5126211
#> 3  -0.5335379 -0.13874092 -1.6966959 -0.5079304 -0.6240419 -1.749320 0.4450619
#> 4  -0.5422436 -0.37299204 -1.7774193 -0.4550787 -0.6202629 -1.528626 0.4248583
#> 5  -0.7096177 -0.37299257 -1.6981375 -0.5351259 -0.9271036 -1.841422 0.3179340
#> 6  -0.5642111 -0.38644846 -1.6620792 -0.2516813 -0.9396287 -1.836708 0.3378222
#> 7  -0.6887217 -0.37443740 -1.7752978 -0.1803544 -1.1168335 -1.836724 0.2726231
#> 8  -0.6172529 -0.38646063 -1.7752611 -0.1696928 -1.3140298 -1.970052 0.3263477
#> 9  -0.6172506 -0.41714033 -1.8366983 -0.2469472 -0.9396292 -1.695244 0.3434938
#> 10 -0.5589418 -0.33904990 -1.6966913 -0.1616829 -1.2832281 -1.876301 0.4381230
#> 11 -0.3080849 -0.03082189 -1.3430481  0.1836984 -0.9477064 -1.461280 0.5714412
#>        grp3_132  grp3_133  grp3_134   grp3_135    grp3_136     grp3_137
#> 1   0.234054464 -2.431048 -1.498320 -0.4942064  0.33971965  0.095115272
#> 2   0.160463943 -2.749902 -1.615169 -0.9049727  0.33421328  0.060727167
#> 3   0.033328855 -2.370568 -1.776353 -1.0596916  0.27503822  0.014470791
#> 4  -0.027667580 -2.371852 -2.120405 -1.0915101  0.08104015  0.019281730
#> 5  -0.037317793 -2.371856 -2.174532 -0.9941518 -0.04254735 -0.006546069
#> 6  -0.083253685 -2.369278 -2.281494 -1.0980121 -0.10194147  0.044544185
#> 7  -0.187664308 -2.369302 -2.430319 -1.0706257 -0.19987405  0.150082665
#> 8  -0.278103342 -2.172234 -2.430257 -1.0572660 -0.29847989  0.090689860
#> 9  -0.262983912 -2.172252 -2.369263 -1.0822307 -0.27809656  0.205139701
#> 10 -0.265146381 -2.057450 -2.370558 -0.9924680 -0.23553419  0.174234510
#> 11  0.006640076 -1.529423 -1.829523 -0.9241661  0.04717021  0.691236561
#>     grp3_138   grp3_139   grp3_140   grp3_141   grp3_142   grp3_143   grp3_144
#> 1  1.2341416  0.5398513 -0.7784940  0.2592604 -0.6622721  0.3063089 -0.8858437
#> 2  1.1511555  0.2426632 -0.3006516  0.2867209 -0.9483876  0.1843256 -0.9924743
#> 3  1.0182415  0.1445471 -0.1671670  0.2867300 -1.1352475  0.1044169 -0.9334065
#> 4  0.8317555 -0.2917804 -0.4732607 -0.1692378 -1.6669116 -0.4155899 -1.0451209
#> 5  0.7362114 -0.3471643 -0.7972510 -0.5164779 -1.8775771 -0.6941436 -1.1199491
#> 6  0.6364647 -0.6172464 -0.4989581 -0.7754701 -2.1722654 -0.7384376 -1.1057534
#> 7  0.5375503 -0.7022732 -0.4444087 -0.8501159 -2.1722864 -0.8927822 -0.9226473
#> 8  0.5447263 -0.8839612 -0.4363984 -0.9226441 -2.3692424 -0.9029003 -0.8725610
#> 9  0.5614395 -0.8228851 -0.3496800 -0.9118288 -2.4302784 -0.8927814 -0.8604707
#> 10 0.6393468 -0.8622841 -0.2849606 -0.7958526 -2.1733908 -0.9132233 -0.8524864
#> 11 0.9569182 -0.4051224 -0.2245450 -0.5411182 -1.7230902 -0.4562622 -0.3896324
#>       grp3_145    grp3_146    grp3_147  grp3_148  grp3_149   grp3_150
#> 1  -0.05970006  0.17424269 -0.04627661 -1.418114 -1.043993  0.1260085
#> 2  -0.07831921  0.19580612 -0.26974339 -1.876312 -1.971300 -0.4060477
#> 3  -0.09637986  0.16740396 -0.38921219 -1.971295 -2.173399 -0.7059537
#> 4  -0.26199578 -0.03731774 -0.58361581 -2.285329 -2.174529 -0.5490636
#> 5  -0.56811876 -0.27139664 -0.62026379 -2.622950 -1.972511 -0.5223891
#> 6  -0.82288324 -0.27250897 -0.76164324 -2.749077 -1.970078 -0.7022742
#> 7  -1.05727448 -0.33055060 -0.94710206 -2.749107 -2.055236 -0.4923816
#> 8  -1.30127473 -0.40420907 -1.02157432 -2.749032 -2.117532 -0.5124152
#> 9  -1.00450593 -0.34254617 -0.99077821 -2.749057 -2.281479 -0.6609179
#> 10 -1.30199798 -0.39465852 -0.92487932 -2.621256 -1.971288 -0.3189024
#> 11 -0.70754444 -0.04781615 -0.75961885 -2.332424 -1.617700 -0.1706536
#>      grp3_151    grp3_152     grp3_153     grp3_154   grp3_155   grp3_156
#> 1   0.2167725 -0.32568890  0.005896701  0.225876588 -0.3530913 -0.4060384
#> 2  -0.4702610 -0.09638836 -0.439063660 -0.010441683 -0.5589455 -0.3332467
#> 3  -0.7646758 -0.24886372 -0.566168109 -0.002117857 -0.6187570 -0.3701531
#> 4  -0.7540239 -0.49021778 -0.516477160 -0.108771361 -0.6941426 -0.4960240
#> 5  -0.6146345 -0.41559047 -0.473261413 -0.267296805 -0.6636318 -0.5490643
#> 6  -0.6777617 -0.34253886 -0.512406050 -0.073752646 -0.6887229 -0.4225911
#> 7  -0.5999550 -0.06291875 -0.538060091 -0.021418715 -0.6226705 -0.3225992
#> 8  -0.6609196 -0.48591992 -0.545101380 -0.056674583 -0.5715022 -0.2831131
#> 9  -0.7944488 -0.15874050 -0.635986875 -0.068539540 -0.6105668 -0.4103231
#> 10 -0.5470861 -0.33324432 -0.646070747  0.053914420 -0.6378391 -0.2488701
#> 11 -0.3194991  0.04638639 -0.035079927  0.366624931 -0.1475909  0.0465006
#>      grp3_157     grp3_158    grp3_159  grp3_160    grp3_161  grp3_162
#> 1  1.10970751 -0.446997956 -0.62403962 0.3597802 -0.02455579 0.6567334
#> 2  0.87432685 -0.248871973  0.05391437 0.7045986 -0.35310037 0.5874692
#> 3  0.73404014  0.024062804  0.27000528 0.9083173 -0.26973517 0.4539258
#> 4  0.73621031  0.163306410  0.50742486 1.0545338 -0.19315357 0.4882231
#> 5  0.65880613 -0.142430657  0.34160796 0.8417514 -0.29721878 0.4470624
#> 6  0.37251880  0.057723852  0.62610559 0.9194455 -0.18767299 0.5029268
#> 7  0.42736152  0.009641911  0.59206071 0.8182353 -0.25167356 0.4735843
#> 8  0.15004371  0.057700873  0.69273331 0.8424528 -0.27252392 0.5028930
#> 9  0.02885022  0.121314035  0.66852256 0.8375753 -0.44441972 0.4831441
#> 10 0.34613816 -0.065642579  0.53084576 0.7246400 -0.48019104 0.4450502
#> 11 0.74701151  0.589961425  1.23911428 1.4463701  0.12552807 0.7524195
#>        grp3_163    grp3_164 grp3_165 grp3_166  grp3_167    grp3_168    grp3_169
#> 1   0.445059144 -0.10536391 1.026236 1.810980 1.4976895  0.90831239 -0.07831094
#> 2  -0.024563897 -0.22686477 1.184490 1.810978 1.0829368  0.45391639  0.53084857
#> 3  -0.431452463 -0.31889650 1.254219 1.875620 0.9587817  0.19581511  0.82056776
#> 4  -0.231008806 -0.29721835 1.370439 1.950280 1.0363799  0.41415978  0.98111045
#> 5  -0.169238006 -0.30892382 1.274846 1.837323 0.9746466  0.60297564  0.70690259
#> 6  -0.317196290 -0.28309842 1.302524 1.948861 1.0280038  0.34350886  0.91944554
#> 7  -0.173985902 -0.26806118 1.403512 1.835823 0.9368019  0.43529022  0.92996318
#> 8  -0.349685830 -0.18769000 1.486127 1.915141 0.8718301  0.38699646  0.93673209
#> 9  -0.550801569 -0.24001439 1.536940 1.948827 0.8468199 -0.01435561  0.92713721
#> 10 -0.424820634 -0.16717400 1.388116 1.836534 0.8871487  0.30063653  0.78265725
#> 11 -0.002649509  0.07484182 1.906514 1.825395 1.7147422  0.67299889  1.37561821
#>        grp3_170    grp3_171  grp3_172   grp3_173   grp3_174  grp3_175
#> 1   0.958776564 -0.75139035 0.5545793 -0.2747937 -0.3762998 0.8743322
#> 2   0.306301797 -0.06564334 1.0939845  0.3162123  0.4298640 0.6567273
#> 3  -0.154192583  0.09511635 1.2727997  0.4979283  0.6864483 0.6701917
#> 4   0.009458147  0.35607522 1.2165734  0.5421616  0.7568970 0.5890583
#> 5   0.211573452  0.22053690 1.0244867  0.2713938  0.5485872 0.7270466
#> 6   0.090713640  0.42106809 1.0097706  0.4091377  0.5103085 0.5520761
#> 7   0.139803161  0.37253586 0.9234969  0.3507663  0.4831795 0.5792016
#> 8  -0.068549053  0.48991778 0.9154549  0.3196408  0.3793488 0.6768005
#> 9  -0.337186277  0.47354918 0.9470342  0.4430310  0.3378073 0.5990033
#> 10 -0.002126136  0.35341046 0.7910090  0.2167642  0.2657341 0.5939548
#> 11  0.259097577  0.82503601 1.5236745  1.1338514  1.0943845 1.0330888
#>       grp3_176  grp3_177    grp3_178  grp3_179 grp3_180 grp3_181 grp3_182
#> 1  -0.84103008 0.7544322  0.67942071 0.5126276 1.254212 1.671097 1.761120
#> 2  -0.54708973 1.0262316  0.25124792 0.7438908 1.502744 1.949563 1.615781
#> 3  -0.17673172 1.1511661  0.30064712 0.6794245 1.475281 1.984310 1.507776
#> 4  -0.26729642 1.1984860  0.38453373 0.8896894 1.697748 2.039401 1.617052
#> 5  -0.22115449 0.9453656  0.19263095 0.5890591 1.380191 1.811791 1.415938
#> 6   0.08163523 0.9697458  0.20515318 0.8280235 1.452093 1.915209 1.403479
#> 7   0.02887364 0.9525780  0.17762522 0.9155239 1.496893 1.760310 1.326435
#> 8   0.40175518 1.0522144  0.06414399 0.9470114 1.486127 1.695315 1.148930
#> 9   0.36432591 1.1073842  0.09070145 1.0166709 1.560635 1.738326 1.189729
#> 10  0.00588902 0.8743220 -0.03998096 0.8966445 1.327228 1.615772 1.214394
#> 11  0.56279414 1.7147422  0.49453152 1.4463701 1.825395 1.825395 1.825395
#>       grp3_183  grp3_184    grp3_185    grp3_186  grp3_187  grp3_188
#> 1   0.11502304 1.5808007  0.32237013 -0.38921100 1.0454449 1.3677370
#> 2   0.29390471 1.0042928  0.12600086  0.04731788 0.7246439 0.9169753
#> 3   0.17424413 0.7624413 -0.03254929 -0.02455527 0.5735552 0.6191700
#> 4  -0.01435769 0.8644569 -0.14946699 -0.08284792 0.7270456 0.7462486
#> 5  -0.35649746 0.8229083 -0.30281354 -0.48338784 0.9145636 0.9145636
#> 6  -0.41713386 0.7893883 -0.31062073 -0.42923129 0.7519583 0.8086448
#> 7  -0.57149158 0.7993043 -0.34253249 -0.37986994 0.7415585 0.8468648
#> 8  -0.53807202 0.6167200 -0.45149076 -0.42924243 0.5920044 0.6462710
#> 9  -0.59996192 0.6260874 -0.58828123 -0.37445024 0.6364463 0.7519386
#> 10 -0.69144143 0.5645147 -0.48807264 -0.54016170 0.7438867 0.8396661
#> 11 -0.26976129 1.0741109 -0.05777547  0.01576773 1.1005631 1.1486453
#>       grp3_189    grp3_190    grp3_191  grp3_192  grp3_193  grp3_194  grp3_195
#> 1  -1.11838598 -0.55893578 -1.55120930 1.4237394 1.0124234 0.9491390 0.6100645
#> 2   0.89664952  1.04544002  0.25925315 1.1351557 0.6701819 0.9721940 0.3397127
#> 3   0.56452733  1.12345967  0.33972188 0.9853935 0.6287749 1.0534029 0.1350973
#> 4   0.19819643  0.76409239 -0.02140937 1.0470911 0.6654458 0.7069016 0.1865232
#> 5  -0.38414195  0.10861447 -0.60707256 1.0297044 0.8044190 0.4929528 0.2774525
#> 6  -0.90290067 -0.41031659 -1.03381341 0.9470560 0.6927717 0.4735657 0.2309528
#> 7  -0.71384244 -0.14305093 -0.81454812 0.9767214 0.6849614 0.5029458 0.1821446
#> 8  -1.00450322 -0.61056920 -0.89278019 0.8845772 0.5989851 0.4352401 0.1095613
#> 9   0.24795951  0.93675472 -0.51241144 0.8718519 0.5272334 0.5029100 0.1990345
#> 10 -0.70595660 -0.02456338 -0.60352619 0.9587661 0.6191570 0.5735425 0.2592520
#> 11 -0.05713175  0.28070815 -0.09667884 1.7147422 0.9972826 0.9761701 0.5492000
#>        grp3_196    grp3_197  grp3_198   grp3_199  grp3_200  grp3_201  grp3_202
#> 1   0.389970599 -0.34485483 0.8018540 -0.4390543 0.2657425 1.1844945 0.2083721
#> 2   0.520210950 -0.68033785 0.8616219 -0.7316737 0.2750291 1.3049779 0.4175138
#> 3   0.389973058 -0.58145069 1.1351662 -0.6803303 0.5466809 1.3049887 0.5874788
#> 4   0.514936442 -0.07330118 1.3418762 -0.6636309 0.6029748 1.2842510 0.9106644
#> 5  -0.056465621 -0.91461288 0.9453656 -1.3731512 0.2465445 0.7170506 0.3364990
#> 6  -0.006485709 -0.80393403 0.8949674 -0.9226450 0.1873538 0.3793796 0.3319303
#> 7  -0.091241730 -0.83919481 0.8376200 -1.0045096 0.2560094 0.4616974 0.4506124
#> 8  -0.251696741 -0.78675030 0.7893476 -1.2282628 0.1990215 0.3042992 0.5102745
#> 9  -0.180372295 -0.55665700 1.0522385 -0.9029017 0.3094603 0.7415164 0.7992607
#> 10 -0.325695603 -0.68033341 0.9169702 -1.2594719 0.3597714 0.5202082 0.5545700
#> 11  0.111067437 -0.14592939 1.5236745 -0.6904119 0.6162629 0.9143842 1.1523701
#>      grp3_203   grp3_204   grp3_205 grp3_206   grp3_207   grp3_208  grp3_209
#> 1  -0.1963787 -1.3152197 -0.6378333 1.487090 0.73403607 0.70460445 0.2512552
#> 2   0.3006380 -0.3946613 -0.8410407 1.630956 0.58133358 0.46575317 0.4450525
#> 3   0.4857007 -0.1308816 -0.8622824 1.630967 0.49145770 0.32838235 0.5202205
#> 4   0.4802818  0.1569241 -1.2185366 1.425396 0.31793353 0.27139343 0.4192538
#> 5   0.5759741 -0.3409034 -1.2840852 1.320067 0.23713949 0.25451950 0.3560757
#> 6   0.8846203 -0.3496728 -1.3534238 1.496860 0.40178653 0.13978958 0.7222450
#> 7   0.8949922 -0.3371725 -1.2823818 1.365062 0.28383854 0.19343156 0.7692202
#> 8   0.8181701 -0.4226023 -1.4161562 1.317360 0.04452153 0.08161167 0.6619537
#> 9   0.8587669 -0.1650979 -1.4625104 1.386291 0.04453252 0.03798017 0.6619728
#> 10  0.9249612 -0.3701586 -1.4181176 1.283188 0.14453773 0.20152248 0.6951474
#> 11  1.3046402  0.1854968 -0.8352823 1.825395 0.53241397 0.53241397 1.3046402
#>      grp3_210    grp3_211 grp3_212     grp3_213 grp3_214  grp3_215  grp3_216
#> 1  -0.2538201  0.45392305 1.836547  0.573551849 2.267901 1.2053375 1.7084929
#> 2  -0.3820206  0.10440808 1.875608  0.546671357 2.267900 1.2542079 1.4655394
#> 3  -0.4942080 -0.06563479 1.810989  0.465762603 2.324549 1.0400805 1.3335960
#> 4  -0.9422554 -0.44171312 1.811788  0.192630678 2.324791 0.8417502 1.1606679
#> 5  -1.0739644 -0.43366576 1.721379 -0.073301289 2.324794 0.6819936 1.1450200
#> 6  -1.0572702 -0.45148017 1.695377 -0.021429885 2.324301 0.7318542 1.0735039
#> 7  -1.1330980 -0.48590594 1.719345  0.002325255 2.324348 0.6620128 1.0384392
#> 8  -1.0980069 -0.59996450 1.670021 -0.027779736 2.324223 0.6073325 0.9271147
#> 9  -1.1693532 -0.52536854 1.785537 -0.056664916 2.267155 0.5858690 0.8997359
#> 10 -1.0068636 -0.53354214 1.786453  0.104407830 2.324523 0.5813305 1.0262258
#> 11 -0.6166933 -0.14012798 1.825395  0.152927112 1.825395 0.9463104 1.8253945
#>     grp3_217   grp3_218    grp3_219    grp3_220   grp3_221  grp3_222 grp3_223
#> 1  0.3114562 -1.1710315 -0.61259960 -1.28322423 -1.2738651 0.6486040 1.318738
#> 2  0.8205577 -0.9585624 -0.66228211 -0.16717537 -0.5661758 1.0262316 1.562235
#> 3  0.3461491 -1.1560295 -0.29509435  0.06791605 -0.4880681 1.0288578 1.594226
#> 4  0.6819926 -1.2840834 -0.32059108  0.23713915 -0.2868040 0.9811104 1.581954
#> 5  1.2165751 -1.6669139 -0.82588525 -0.25595247 -0.2868045 0.8896907 1.539696
#> 6  1.0924953 -1.4437769 -0.85011466 -0.06852915  0.2559941 1.1642283 1.785569
#> 7  1.3174443 -1.5381103 -0.73843711 -0.02774776  0.4018040 1.2519501 1.915249
#> 8  1.5069455 -1.4770157 -0.76164629  0.03796927  0.2911168 1.3024711 1.874508
#> 9  1.3391113 -1.4375112 -0.48591585  0.18211714  0.3196554 1.4034521 1.874543
#> 10 1.5383168 -1.4081066 -0.80638252 -0.01044125  0.1795385 1.3677254 1.916149
#> 11 1.5980358 -1.1053684 -0.01865945  0.48956696  0.6432462 1.7147422 1.906514
#>     grp3_224  grp3_225    grp3_226  grp3_227 grp3_228   grp3_229    grp3_230
#> 1  1.7084929 0.6191665 -0.73166342 0.9169806 1.562238 1.06813155 -0.06563511
#> 2  1.4655394 0.7713147 -0.08695255 0.9491338 1.761117 0.95877145  0.13508838
#> 3  1.3335960 0.8102697  0.12600973 0.9169855 1.786475 0.82989657  0.36748942
#> 4  1.1606679 0.6975335  0.34877132 0.8229071 1.539694 0.53444237  0.10861432
#> 5  1.1450200 0.7568980 -0.24616411 0.7853197 1.307437 0.41416038 -0.40255799
#> 6  1.0735039 1.0214272 -0.33717891 1.1074079 1.339138 0.35767383 -0.88396113
#> 7  1.0384392 1.1336478 -0.41712861 1.1642572 1.189783 0.29116128 -0.79444698
#> 8  0.9271147 1.0657341 -0.49896749 1.2820982 1.114917 0.18210436 -0.81454920
#> 9  0.8997359 1.0808063 -0.12740595 1.2707146 1.121612 0.08162314 -0.54509809
#> 10 1.0262258 1.1096966 -0.14627762 1.4237275 1.234130 0.26999492 -0.77849935
#> 11 1.8253945 1.5980358  0.16879393 1.7147422 1.825395 0.65357974 -0.40165251
#>       grp3_231 grp3_232    grp3_233   grp3_234 grp3_235  grp3_236 grp3_237
#> 1  -1.21704389 1.340506 -0.33323764 -1.2594679 1.454778 0.7624371 1.966918
#> 2  -0.48019428 1.650550 -0.04628478 -0.7785044 1.720336 0.9343724 1.984298
#> 3  -0.20999101 1.615793  0.26574441 -0.5275776 1.761128 0.8923234 1.836556
#> 4  -0.04984548 1.476868 -0.09061823 -0.5097953 1.498526 0.5001892 1.837320
#> 5  -0.44957547 1.415938 -0.65824419 -0.9836292 1.380191 0.3845343 1.617054
#> 6  -0.37444329 1.560664 -0.44441356 -0.9806860 1.486185 0.4831608 1.719308
#> 7  -0.34966650 1.548574 -0.52535966 -0.8228841 1.486218 0.4091553 1.670119
#> 8  -0.19391988 1.518087 -0.49239531 -0.8039360 1.403423 0.2559663 1.579593
#> 9  -0.05002089 1.630106 -0.39319597 -0.5188176 1.452066 0.2784265 1.614493
#> 10 -0.22686304 1.594205 -0.40032576 -0.6702181 1.518957 0.4298618 1.696547
#> 11  0.25874810 1.825395 -0.04639359 -0.3242444 1.825395 0.7882350 1.825395
#>     grp3_238   grp3_239 grp3_240 grp3_241  grp3_242 grp3_243 grp3_244 grp3_245
#> 1  1.5189687 -0.2898209 1.538329 1.916163 0.3283802 2.038572 1.123454 2.497698
#> 2  1.3232173 -0.4003285 1.696557 2.153279 0.3461399 2.082790 1.423736 2.497697
#> 3  1.0681372 -0.5401575 1.650561 2.038583 0.3972279 2.153291 1.538337 2.497710
#> 4  0.8081498 -0.9347690 1.761975 1.917126 0.2545191 1.984951 1.563814 2.497806
#> 5  0.7734377 -1.2949934 1.684935 1.984954 0.1981967 1.651595 1.177428 2.497809
#> 6  0.7607772 -0.9320399 1.835784 2.152624 0.4899513 1.326404 1.052262 2.497609
#> 7  0.7318766 -1.0822364 1.983701 2.267238 0.3435255 1.290901 1.092523 2.497658
#> 8  0.7414964 -1.1922379 2.037684 2.267115 0.5272162 1.212183 1.025334 2.497527
#> 9  0.7222256 -1.0338129 2.082202 2.324264 0.4157634 1.317387 1.365004 2.497570
#> 10 0.7624272 -1.2452449 2.082778 2.267886 0.4769228 1.166104 1.151149 2.497683
#> 11 1.0925375 -0.6016844 1.825395 1.825395 0.8237635 1.825395 1.825395 1.825395
#>    grp3_246   grp3_247     grp3_248  grp3_249  grp3_250   grp3_251  grp3_252
#> 1  2.324538 -0.5079287  1.205337489 1.1511601 0.9381450 0.38196312 0.3342203
#> 2  2.324537 -0.7646832  1.068126762 1.1097029 1.0042928 0.37488350 0.3223631
#> 3  2.267912 -0.7958504  0.924976539 1.0829472 0.9722042 0.21677416 0.3597825
#> 4  2.268623 -1.2752647  0.457235694 0.6718292 0.6422358 0.05740585 0.3364985
#> 5  2.083362 -1.5286282  0.120430204 0.5421624 0.4192544 0.48822375 0.3031692
#> 6  2.267192 -1.2823741  0.268605810 0.7691972 0.6685413 0.46167900 0.5447611
#> 7  2.082279 -1.1922516 -0.006474316 0.7800292 0.5447808 0.66856275 0.5990417
#> 8  2.082164 -1.1534761 -0.062949341 0.7519184 0.4956177 0.89971375 0.5520411
#> 9  2.037721 -1.1922430  0.109573153 0.7799861 0.4899346 0.73183475 0.5447438
#> 10 2.038559 -0.9821601 -0.053252645 0.8298819 0.4116460 0.84861008 0.6009924
#> 11 1.825395 -0.7318246  0.491316964 0.9463104 0.9175439 1.28934650 1.0912136
#>     grp3_253    grp3_254    grp3_255     grp3_256    grp3_257    grp3_258
#> 1  0.1843331 -0.15419249 -0.03997341  0.429870747  1.50776879  0.28672802
#> 2  0.3597732 -0.28012614 -0.16168424  0.389963752  1.16611053  0.20152332
#> 3  0.4381346 -0.31338613 -0.14627044  0.417523186  1.23414804  0.31622145
#> 4  0.3995319 -0.38414141 -0.50372455  0.071641199  1.07047584  0.43238568
#> 5  0.4068406 -0.23878885 -0.59390369 -0.082848040  0.50018988  0.05740593
#> 6  0.5858868 -0.13503514 -0.61056248  0.002313729  0.02886171  0.01949571
#> 7  0.5614768 -0.08324344 -0.64387481  0.038003829 -0.01433332 -0.03737753
#> 8  0.5791458 -0.12741469 -0.68872769 -0.073772426 -0.26299089 -0.17401213
#> 9  0.5520587 -0.20726672 -0.92264577 -0.062939756 -0.10195149 -0.03739920
#> 10 0.6287618  0.01446239 -0.61260546  0.067907234 -0.15419970 -0.05970754
#> 11 1.1676637  0.19490516 -0.30878715  0.217359120  0.53609628  0.13399428
#>    grp3_259   grp3_260   grp3_261   grp3_262   grp3_263     grp3_264   grp3_265
#> 1  2.153280 0.60100187 -0.2175720 -0.9334028 -2.0574484 -0.394652020 -1.1067186
#> 2  1.916161 0.67941474 -0.2898299 -0.8943178 -1.5269502 -0.361993175 -1.1710431
#> 3  1.696568 0.69516103 -0.1053638 -0.8321354 -1.2452462 -0.406039711 -1.0439969
#> 4  1.457459 0.29666362 -0.3651033 -1.0137499 -1.1946440 -0.658243250 -0.8958359
#> 5  1.476870 0.03776846 -0.9496592 -0.8428675 -1.3559737 -0.682889415 -1.2608693
#> 6  1.386318 0.12132657 -1.4961731 -0.8604692 -1.1330926 -0.393189241 -1.2929642
#> 7  1.339169 0.05773620 -1.3704975 -0.6680504 -1.0980170 -0.386442746 -1.1534884
#> 8  1.231760 0.05039515 -1.2155482 -0.5882840 -1.0706167 -0.246954347 -1.4625017
#> 9  1.231786 0.01948432 -1.1168269 -0.4769776 -1.0572694 -0.187682045 -1.2823708
#> 10 1.340494 0.15349811 -1.0232884 -0.5013505 -1.2170480 -0.096387417 -1.1828413
#> 11 1.825395 0.26511067 -0.8150444 -0.1541515 -0.8489838 -0.004181526 -0.8380745
#>     grp3_266  grp3_267  grp3_268 grp3_269    grp3_270   grp3_271    grp3_272
#> 1  1.3272395 0.7910189 0.8966548 1.594218 -0.09638004 -0.5335360  0.31621934
#> 2  1.2727889 0.6191603 0.6287652 1.518965 -0.52061123 -0.8321427 -0.03998157
#> 3  0.9969823 0.5398544 0.5051839 1.283206 -0.64606738 -0.9049655 -0.35309228
#> 4  1.1367004 0.6127485 0.5225313 1.236468 -0.52238834 -0.9941504 -0.30892338
#> 5  1.0470926 0.7462497 0.6879457 1.328082 -0.33590958 -0.9070238 -0.19315384
#> 6  0.9367765 0.6463083 0.5711201 1.251920 -0.45818523 -1.0045061 -0.30421156
#> 7  0.9525780 0.6364857 0.6167772 1.212264 -0.42258588 -0.9118301 -0.15028032
#> 8  0.8587453 0.5374962 0.4157476 1.164178 -0.56421896 -1.0338098 -0.30422573
#> 9  0.9271372 0.6073508 0.4616626 1.212209 -0.62267686 -0.9470999 -0.31720388
#> 10 0.9853777 0.6567238 0.4979162 1.197032 -0.52060774 -0.9722973 -0.24309778
#> 11 1.7147422 0.9607958 0.8606621 1.825395  0.04463164 -0.6103739  0.13751798
#>       grp3_273  grp3_274     grp3_275    grp3_276  grp3_277   grp3_278
#> 1   0.47693185 1.3049820 -0.400319273 -0.07092005 0.1572718 -1.2452410
#> 2   0.08584081 0.9083071 -0.533545700  0.16739502 0.4079784 -0.7514007
#> 3  -0.19637895 0.6393600 -0.514445388  0.51263065 0.7959563 -0.4129586
#> 4  -0.17327836 0.6588052 -0.568117946  0.63142584 0.9224646 -0.2387885
#> 5   0.06372828 0.8317567 -0.455079382  0.93199866 1.0902673  0.1387547
#> 6  -0.15028956 0.6849396 -0.579277245  0.95716255 1.2031416  0.3196702
#> 7  -0.10193151 0.7121442 -0.564208058  0.97672139 1.1336478  0.3793968
#> 8  -0.21399119 0.5858510 -0.518821205  1.02533436 1.1897033  0.4616462
#> 9  -0.25168962 0.6462899 -0.628811449  0.96972368 1.1405929  0.4352561
#> 10 -0.13088894 0.7045947 -0.527581836  0.97218851 1.1294187  0.4175117
#> 11  0.14501037 1.0154847  0.002978739  1.35682079 1.5180294  0.9137563
#>     grp3_279   grp3_280    grp3_281  grp3_282  grp3_283    grp3_284   grp3_285
#> 1  0.5645240 -0.6306703 -0.23552722 0.6393564 0.6864444  0.40431948 -0.7157437
#> 2  0.8102597 -0.1714883 -0.24309961 1.2341373 0.5398449 -0.29510253 -0.5144532
#> 3  1.2144117  0.1900012 -0.34485582 1.4924204 0.7713246 -0.08694403  0.1843345
#> 4  1.2748438  0.2896178 -0.35649695 1.4880003 0.9603879  0.02860178  0.0443151
#> 5  1.5085427  0.5834902 -0.29178087 1.6849350 1.1186375  0.36187647  0.5149372
#> 6  1.5369681  0.8280235 -0.40419733 1.6301355 1.1336194  0.36434118  0.3094750
#> 7  1.4737143  0.8086684 -0.28784866 1.5070377 1.0280306  0.31451247  0.2981240
#> 8  1.4861270  0.8279818 -0.23226880 1.5606047 0.9831498  0.27257932  0.3576436
#> 9  1.5181172  0.7893683 -0.27251707 1.4968316 1.0166709  0.29809352  0.2635228
#> 10 1.4547664  0.7713105 -0.26974141 1.5622255 0.9491285  0.30630033  0.3162107
#> 11 1.9065140  1.4886727  0.05707274 1.9065140 1.5980358  0.59544454  0.7666423
#>     grp3_286  grp3_287  grp3_288 grp3_289  grp3_290  grp3_291    grp3_292
#> 1  0.3972254 0.6701879 0.5466777 1.475273 1.0262365 0.2939118  0.49792536
#> 2  0.2083647 0.5545729 0.7910133 1.580797 0.6637155 0.8298865 -0.25953182
#> 3  0.2803616 0.5545825 1.0939950 1.739137 0.8486248 1.1750483 -0.32568979
#> 4  0.2774521 0.4698204 1.3200647 1.787371 1.0244852 1.2565007 -0.39611282
#> 5  0.2822687 0.4572363 1.4985278 1.917129 1.1186375 1.4880024 -0.04984555
#> 6  0.1500689 0.3949076 1.5070037 1.874575 1.0437966 1.5181455 -0.05001029
#> 7  0.1714238 0.3949250 1.6495554 1.874615 1.1074359 1.6301714 -0.06851868
#> 8  0.1575949 0.4210363 1.7602080 1.948792 1.1149172 1.7192448 -0.11920078
#> 9  0.1934038 0.4273277 1.6494895 1.915176 1.0657584 1.6700529 -0.21398358
#> 10 0.2750278 0.5051717 1.6710840 1.949551 1.0829307 1.6505403 -0.07092751
#> 11 0.5323222 0.7524195 1.9119526 1.825395 1.6408901 1.9119526  0.27797784
#>     grp3_293  grp3_294  grp3_295 grp3_296 grp3_297     grp3_298 grp3_299
#> 1  0.9249716 0.7713204 0.0679151 1.283199 1.272793 -0.764672845 1.739128
#> 2  1.3232173 0.6393503 0.4856912 1.340502 1.404973 -0.824393409 1.836545
#> 3  1.3881354 0.6567371 0.7145945 1.404984 1.492420 -0.722738344 1.949574
#> 4  1.5085406 0.5759732 1.0850491 1.519798 1.631790 -0.347163824 2.083359
#> 5  1.5950345 0.6215674 1.2565025 1.631792 1.787374 -0.158078511 2.153952
#> 6  1.5934102 0.5614570 1.2821510 1.579653 1.738357 -0.143060282 2.082236
#> 7  1.5934455 0.5711401 1.4521266 1.695414 1.738394  0.131430030 2.152668
#> 8  1.6300743 0.6260688 1.5933499 1.738293 1.835717  0.193390898 2.152550
#> 9  1.5796239 0.5920225 1.4861568 1.719278 1.835752  0.002302536 2.152589
#> 10 1.5807883 0.6485944 1.4870776 1.720326 1.810967  0.167394384 2.153266
#> 11 1.9065140 0.9192235 1.5980358 1.906514 1.906514  0.367857412 1.825395
#>    grp3_300   grp3_301  grp3_302     grp3_303  grp3_304    grp3_305    grp3_306
#> 1  1.786466 -0.1714797 0.5051809 -0.006514733 0.3674871  0.28035968  0.88715896
#> 2  1.594214 -0.2538289 0.4381253 -0.421519684 1.0353157 -0.07092830  0.04115503
#> 3  1.671105 -0.1836273 0.3748928 -0.418790438 1.4547857  0.08584961 -0.07091975
#> 4  1.721377  0.1204300 0.5570715 -0.250781862 1.3899383  0.09949178 -0.14243045
#> 5  1.876657  0.2297004 0.6975344 -0.198845446 1.5819559  0.43238630  0.13066428
#> 6  1.810181  0.2479735 0.6546538 -0.268068668 1.6145221  0.41577929  0.03799177
#> 7  1.810220  0.3643581 0.6927936 -0.109436889 1.5485738  0.38704473  0.08164794
#> 8  1.810116  0.3948765 0.7318148  0.002292099 1.5369091  0.30944586 -0.05003065
#> 9  1.810150  0.3319154 0.5791637 -0.292970282 1.5069755  0.29113099 -0.15029903
#> 10 1.875597  0.3899618 0.6794110 -0.217578995 1.4976775  0.32236152  0.03332031
#> 11 1.825395  0.8459196 1.1925744  0.124892087 1.9065140  0.64293600  0.19203037
#>        grp3_307   grp3_308  grp3_309  grp3_310  grp3_311     grp3_312  grp3_313
#> 1  -0.431451074 -0.8063772 0.4175206 0.8616274 0.4229676 -0.086944255 0.5308550
#> 2  -0.461022048 -0.3133943 0.7340303 1.5027445 0.8018484 -0.002126518 0.4914482
#> 3  -0.480186410 -0.1229993 0.9343826 1.5622458 1.1970497  0.381965544 0.4229703
#> 4  -0.535125165  0.3082793 1.1120021 1.6515929 1.1868538  0.406840009 0.3706169
#> 5  -0.529784077  0.5074256 1.1984877 1.7619777 1.2842528  0.650891585 0.3706174
#> 6  -0.556652105  0.5272506 1.1821306 1.6700829 1.1489797  0.450594143 0.2981080
#> 7  -0.518808517  0.7608001 1.1897831 1.5796880 1.0658090  0.415796997 0.3319468
#> 8  -0.467242894  0.8086036 1.2518678 1.4520365 0.9571177  0.247945868 0.3507194
#> 9  -0.506064341  0.8086245 1.1642040 1.4736530 0.9950671  0.212974118 0.3725034
#> 10 -0.470257814  0.8205532 1.1844834 1.4752613 0.9969664  0.311447622 0.3972166
#> 11  0.002978739  1.4719165 1.6408901 1.8253945 1.6408901  0.981019351 0.7336934
#>       grp3_314  grp3_315  grp3_316     grp3_317 grp3_318    grp3_319   grp3_320
#> 1  -0.56616609 0.4381320 0.6287713 -1.293973734 1.388128  0.05392224 -0.8751945
#> 2  -0.26514834 0.4979188 0.7624314 -1.354706229 1.538326 -0.05325333 -1.0838231
#> 3  -0.07831067 0.6100679 0.9343826 -1.371821050 1.580809 -0.28982167 -1.2738704
#> 4   0.33037293 0.8503953 1.1252698 -0.914611578 1.672117 -0.11857472 -1.1585656
#> 5   0.48028245 0.9988621 1.2364698 -0.877825579 1.739910 -0.10877152 -1.1076923
#> 6   0.53753082 0.9059532 1.2318111 -0.830525123 1.649519 -0.15873112 -1.2024682
#> 7   0.88464499 1.0522888 1.2909005 -0.498954027 1.614558 -0.04999955 -1.0338174
#> 8   0.99504368 1.1335702 1.3862629 -0.458195662 1.614461 -0.04271189 -0.9567515
#> 9   0.82355506 1.0437737 1.2821254 -0.564215920 1.593381 -0.14306984 -1.0898022
#> 10  0.90830205 1.0939783 1.2727817 -0.514449758 1.630947 -0.01789279 -0.9483816
#> 11  1.48867266 1.5980358 1.7147422 -0.005761769 1.825395  0.24103981 -0.6469331
#>     grp3_321  grp3_322 grp3_323   grp3_324     grp3_325    grp3_326   grp3_327
#> 1  -1.006859 -1.083812 1.166115 0.06073502  0.041162941 -0.25952300 -0.8622790
#> 2  -1.418126 -1.696702 1.388124 0.02405410  0.005888686 -0.21758067 -1.0356912
#> 3  -1.526944 -2.057457 1.423747 0.04732662  0.060735932 -0.21757236 -1.0356842
#> 4  -1.616515 -1.931274 1.595032 0.05009047 -0.101484018  0.09003773 -0.8640889
#> 5  -1.590464 -1.931276 1.563816 0.14927567 -0.062732439  0.18652342 -0.7907746
#> 6  -1.695252 -1.775283 1.473681 0.10019320 -0.127396216  0.22203213 -0.7138433
#> 7  -1.586398 -1.613818 1.518180 0.10020619 -0.207249497  0.32639354 -0.5882740
#> 8  -1.586369 -1.336241 1.422028 0.13976461 -0.379888084  0.27841246 -0.5792847
#> 9  -1.662072 -1.549827 1.422057 0.10018090 -0.193912004  0.22201846 -0.6680555
#> 10 -1.354698 -1.664496 1.507757 0.13508794 -0.209997706  0.28671951 -0.5736485
#> 11 -1.108497 -1.438727 1.825395 0.46771075  0.073938568  0.79842814 -0.1347640
#>      grp3_328  grp3_329  grp3_330 grp3_331 grp3_332     grp3_333    grp3_334
#> 1  -0.9585514 -1.776345 -1.391065 1.630959 1.875611 -0.176731526  0.13509600
#> 2  -1.2832362 -2.370572 -2.431063 2.038571 1.786463  0.095107528  0.22586924
#> 3  -1.3381974 -2.621268 -2.749897 2.082802 1.916172  0.208373678  0.33422249
#> 4  -1.3027135 -2.431820 -2.750715 2.153949 1.876654  0.002222959  0.30316876
#> 5  -1.2752665 -2.431824 -2.750719 2.268626 2.039404  0.019281758  0.28961821
#> 6  -1.2436123 -2.619582 -2.930344 2.037755 1.983660 -0.169675377  0.10958556
#> 7  -1.1057585 -2.281517 -2.619610 2.037797 1.948901 -0.292955307  0.01950749
#> 8  -1.0897983 -2.281460 -2.619540 1.983590 1.785504 -0.398086906 -0.02145094
#> 9  -1.2024658 -2.619564 -2.325451 1.983627 1.760241 -0.322613347 -0.13504478
#> 10 -1.1560289 -2.283406 -2.431048 1.984286 1.739115 -0.306578002 -0.03255733
#> 11 -0.7309339 -1.850386 -1.850386 1.825395 1.825395  0.061962939  0.21735912
#>       grp3_335    grp3_336    grp3_337    grp3_338    grp3_339  grp3_340
#> 1  -0.18362704 -0.14627039 -0.69143580 -0.30657124 -0.50134430 0.5874754
#> 2  -0.27480261 -0.10537227 -0.97230352 -0.19042721 -0.44700735 0.7145846
#> 3  -0.30064343 -0.01788463 -0.67021487 -0.05324476 -0.25952359 0.7246538
#> 4  -0.32876073 -0.16460836 -0.52978332  0.17705956 -0.04254729 0.8081498
#> 5  -0.18316777 -0.10148416 -0.36510383  0.26793663  0.05009055 0.9264705
#> 6  -0.23225288 -0.22268535 -0.35885358  0.27260753  0.07221202 0.8468407
#> 7  -0.35884745 -0.22267719 -0.40419188  0.27845646  0.10959868 0.8280474
#> 8  -0.53195411 -0.33719228 -0.41714529  0.23874831  0.02883944 0.7607372
#> 9  -0.37988263 -0.42923762 -0.49239133  0.17759791 -0.02776969 0.6768197
#> 10 -0.34486148 -0.30064943 -0.41296374  0.23404611  0.06072717 0.7340262
#> 11  0.04387527  0.03249116  0.07750065  0.67538197  0.61803042 1.1177077
#>       grp3_341    grp3_342 grp3_343     grp3_344   grp3_345   grp3_346
#> 1  -0.31889564 -0.30064266 2.082791 -0.032549765 -0.4532811 -0.2488632
#> 2  -0.19638729 -0.17674010 1.739125 -0.130890089 -0.6035302 -0.6187647
#> 3  -0.23552770 -0.27479440 1.720348  0.005897360 -0.6557304 -0.4532826
#> 4  -0.17945756 -0.40787565 1.739908 -0.126799707 -0.7176427 -0.3919646
#> 5  -0.12679989 -0.44171375 1.950282  0.002222962 -0.9422568 -0.1831678
#> 6  -0.04269136 -0.54509307 1.760272 -0.180363162 -0.9118279 -0.1739948
#> 7  -0.05664374 -0.55079336 1.785607 -0.246930981 -0.9907815 -0.2629682
#> 8  -0.14307838 -0.62881365 1.649458 -0.207274421 -0.9396274 -0.3305707
#> 9  -0.19989146 -0.80393613 1.695347 -0.467238585 -0.9806861 -0.3042193
#> 10 -0.12300669 -0.61876061 1.761107 -0.295100393 -1.1352471 -0.3133920
#> 11  0.22573879 -0.02739639 1.825395  0.046364518 -0.4728469  0.0999768
#>       grp3_347    grp3_348   grp3_349  grp3_350   grp3_351   grp3_352
#> 1   0.02406206 -0.05324514 -0.4880664 0.7246498 -0.2697345 -0.1308817
#> 2  -0.57365235 -0.49421589 -0.9334137 0.4229609 -0.5911034 -0.4129666
#> 3  -0.54708197 -0.43905572 -0.7403626 0.3534215 -0.6622745 -0.5589378
#> 4  -0.61463358 -0.27139625 -0.6482496 0.4409778 -0.5757865 -0.7096167
#> 5  -0.32059154 -0.20634828 -0.4204422 0.5344431 -0.5037253 -0.7176437
#> 6  -0.43638742 -0.25704161 -0.4769718 0.3507495 -0.5714945 -0.9567532
#> 7  -0.45147542 -0.23224486 -0.5060550 0.4430653 -0.5450897 -0.7867481
#> 8  -0.39320125 -0.36731499 -0.7138475 0.1775852 -0.8391949 -1.0428662
#> 9  -0.57149925 -0.40420394 -0.7487544 0.1873405 -0.8725620 -1.1814978
#> 10 -0.41879554 -0.36199064 -0.7157493 0.1958053 -0.7316689 -0.9585563
#> 11  0.04705544  0.09392554 -0.2205500 0.5795671 -0.1311471 -0.5843028
#>       grp3_353   grp3_354 grp3_355   grp3_356    grp3_357   grp3_358   grp3_359
#> 1  -0.28495375 -0.8166057 1.966918 -0.3133853  0.52021740 -0.5275757 -0.4248142
#> 2  -0.23553597 -1.2594798 1.487086 -0.9821664 -0.05970827 -1.3152317 -1.0130269
#> 3   0.07663979 -1.2036007 1.367744 -1.0130199 -0.25382067 -1.4389117 -1.2295784
#> 4   0.26250906 -1.0739629 1.328080 -1.0092104 -0.30281311 -1.5004705 -1.3401355
#> 5   0.32505652 -1.0249937 1.519801 -1.0092118 -0.25078222 -1.3925100 -1.3401374
#> 6   0.30432815 -1.0898033 1.365030 -1.0706212 -0.26297571 -1.4161690 -1.5498332
#> 7   0.30434419 -1.2282780 1.386350 -0.9320423 -0.23999803 -1.4625261 -1.3896389
#> 8   0.33779245 -1.2929543 1.326350 -0.8501156 -0.32261954 -1.3534125 -1.6952320
#> 9   0.31448157 -1.3012813 1.189729 -1.0122889 -0.43639367 -1.6137986 -1.9278624
#> 10  0.29390332 -1.1183903 1.318726 -0.9334077 -0.27480060 -1.3381955 -1.3718189
#> 11  0.77086969 -0.7917090 1.825395 -0.6430993  0.06346406 -1.0309703 -1.0208788
#>       grp3_360  grp3_361   grp3_362   grp3_363 grp3_364  grp3_365    grp3_366
#> 1  -0.36198404 0.1572718  0.1674025 -0.2268560 1.650553 0.3748904 -0.26513950
#> 2  -0.30658021 0.4079784 -0.1462789  0.3819563 1.283195 0.5735456 -0.34486392
#> 3  -0.30657205 0.7959563  0.1604729  0.6637253 1.175048 0.7439007 -0.38201258
#> 4  -0.27707109 0.9224646  0.2297000  0.9512167 1.307435 0.9988607 -0.42044155
#> 5  -0.26199615 1.0902673  0.5225320  1.1774283 1.457461 1.0545353 -0.14946720
#> 6  -0.24000593 1.2031416  0.6768385  1.2707401 1.422084 1.0342622 -0.32260582
#> 7  -0.15872201 1.1336478  0.4899701  1.2707706 1.422117 1.0214540 -0.31718954
#> 8  -0.13505343 1.1897033  0.4273117  1.3390834 1.364975 1.0924472 -0.44442433
#> 9  -0.17400400 1.1405929  0.4091219  1.3024985 1.326378 0.9831730 -0.45819123
#> 10 -0.10537127 1.1294187  0.4043106  1.3049705 1.404965 1.0496552 -0.44700430
#> 11  0.08677933 1.5180294  1.1257491  1.5180294 1.825395 1.5980358  0.05707274
#>       grp3_367     grp3_368    grp3_369
#> 1  -0.24309084 -0.006514733 -0.54708004
#> 2  -0.20999933 -0.421519684 -0.71575394
#> 3  -0.19041886 -0.400320517 -0.87519797
#> 4   0.03776841 -0.068329415 -0.80881681
#> 5   0.20399993 -0.164608597 -0.73386928
#> 6   0.19904791  0.131416575 -0.65320730
#> 7   0.30949108  0.319686436 -0.39318363
#> 8   0.21296093  0.283794323 -0.16510619
#> 9   0.13977684  0.157607362 -0.57928188
#> 10  0.25124679  0.225868259 -0.55201007
#> 11  0.79842814  0.798428139 -0.03538885